# docs/swagger.yaml
basePath: /api
definitions:
  domain.AuthRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  domain.BookStoreRequest:
    properties:
      author:
        type: string
      description:
        type: string
      isbn:
        type: string
      language:
        type: string
      pages:
        type: integer
      price:
        type: integer
      published_at:
        type: string
      stock:
        type: integer
      title:
        type: string
    required:
    - author
    - description
    - isbn
    - language
    - pages
    - price
    - published_at
    - stock
    - title
    type: object
  domain.BookUpdateRequest:
    properties:
      author:
        type: string
      description:
        type: string
      isbn:
        type: string
      language:
        type: string
      pages:
        type: integer
      price:
        type: integer
      published_at:
        type: string
      stock:
        type: integer
      title:
        type: string
    type: object
...
  domain.Error:
    properties:
      code:
        type: integer
      errors:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
  domain.Success:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  title: Book Store API Documentation
  version: "1.0"
paths:
  /auth/token:
    post:
      consumes:
      - application/json
      description: Get JWT Token
      parameters:
      - description: user credential
        in: body
        name: auth
        required: true
        schema:
          $ref: '#/definitions/domain.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: token detail
          schema:
            $ref: '#/definitions/domain.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Get JWT Token
      tags:
      - auth
  /books:
    get:
      consumes:
      - application/json
      description: Get list of books
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Size of page (default 10)
        in: query
        name: size
        type: integer
      - description: Search query
        in: query
        name: q
        type: string
      - description: Filter by (title, author, price)
        in: query
        name: filterBy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of books
          schema:
            items:
              $ref: '#/definitions/domain.Success'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Get list of book
      tags:
      - books
    post:
      consumes:
      - application/json
      description: Store book
      parameters:
      - description: book data
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/domain.BookStoreRequest'
      produces:
      - application/json
      responses:
        "201":
          description: book detail
          schema:
            $ref: '#/definitions/domain.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      security:
      - Bearer: []
      summary: Store book
      tags:
      - books
  /books/{id}:
    delete:
      consumes:
      - application/json
      description: Delete book
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success delete book
          schema:
            $ref: '#/definitions/domain.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      security:
      - Bearer: []
      summary: Delete book
      tags:
      - books
    get:
      consumes:
      - application/json
      description: Get book by id
      parameters:
      - description: book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: book detail
          schema:
            $ref: '#/definitions/domain.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Get book by id
      tags:
      - books
    put:
      consumes:
      - application/json
      description: Update book
      parameters:
      - description: book ID
        in: path
        name: id
        required: true
        type: integer
      - description: book data
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/domain.BookUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: book detail
          schema:
            $ref: '#/definitions/domain.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      security:
      - Bearer: []
      summary: Update book
      tags:
      - books
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"